+++ src/knx/bau07B0.h
@@ -18,7 +18,7 @@ class Bau07B0 : public BauSystemBDevice, public ITpUartCallBacks, public DataLin
     
   protected:
     InterfaceObject* getInterfaceObject(uint8_t idx);
-    InterfaceObject* getInterfaceObject(ObjectType objectType, uint8_t objectInstance);
+    InterfaceObject* getInterfaceObject(ObjectType objectType, uint16_t objectInstance);
 
     // For TP1 only
     TPAckType isAckRequired(uint16_t address, bool isGrpAddr) override;
