+++ src/knx/bau091A.cpp
@@ -99,7 +99,7 @@ InterfaceObject* Bau091A::getInterfaceObject(uint8_t idx)
     }
 }
 
-InterfaceObject* Bau091A::getInterfaceObject(ObjectType objectType, uint8_t objectInstance)
+InterfaceObject* Bau091A::getInterfaceObject(ObjectType objectType, uint16_t objectInstance)
 {
     // We do not use it right now. 
     // Required for coupler mode as there are multiple router objects for example
