+++ src/knx/bau091A.h
@@ -21,7 +21,7 @@ class Bau091A : public BauSystemBCoupler, public ITpUartCallBacks, public DataLi
 
   protected:
     InterfaceObject* getInterfaceObject(uint8_t idx);
-    InterfaceObject* getInterfaceObject(ObjectType objectType, uint8_t objectInstance);
+    InterfaceObject* getInterfaceObject(ObjectType objectType, uint16_t objectInstance);
 
     // For TP1 only
     TPAckType isAckRequired(uint16_t address, bool isGrpAddr) override;
