+++ src/knx/bau2920.h
@@ -24,7 +24,7 @@ class Bau2920 : public BauSystemBCoupler
 
   protected:
     InterfaceObject* getInterfaceObject(uint8_t idx);
-    InterfaceObject* getInterfaceObject(ObjectType objectType, uint8_t objectInstance);
+    InterfaceObject* getInterfaceObject(ObjectType objectType, uint16_t objectInstance);
 
     void doMasterReset(EraseCode eraseCode, uint8_t channel) override;
   private:
